requirements.txt
opencv-python
matplotlib
numpy
python-telegram-bot
requests